Player Class, Player Initialization, Card Rework
I created the Player class and began to implement it into the driver. My
next commit will be documenting my changes. Essentially, the Driver now
lets the player select a character to play as, selects the AI opponents,
deals them cards from a shuffled deck, and draws a card from the top of
the deck into play.