Discard method Edited
I edited the discard method to use number 1 through the number of the
player's cards. In addition to this, I tried out a new listing system
for printing out the card's in the player's hand. The cards do not yet
get discarded from the player's hand after discarding them. We can see
if this works better than the earlier version.